+++ src/query/products_components/category_query.rs
@@ -3,6 +3,7 @@ use crate::error::api_error::ApiError;
 use rocket::serde::json::Json;
 use rocket::State;
 use sqlx::{PgPool, Row};
+use crate::data::user_components::claims::Claims;
 
 #[post("/category", data = "<category_data>")]
 pub async fn create_category(
@@ -89,7 +90,8 @@ pub async fn update_category_name(
     Ok("Category successfully updated".to_string())
 }
 #[delete("/category/<id>")]
-pub async fn delete_category_by_id(db_pool: &State<PgPool>, id: i32) -> Result<String, ApiError> {
+pub async fn delete_category_by_id(db_pool: &State<PgPool>, id: i32, claims: Claims) -> Result<String, ApiError> {
+    claims.check_admin()?;
     sqlx::query(
         r#"
         DELETE FROM categories
